+++ app/src/main/java/com/senya/novuswidget/domain/model/ShopItemSP.kt
@@ -1,0 +1,7 @@
+package com.senya.novuswidget.domain.model
+
+data class ShopItemSP(
+    val path: String,
+    val title: String = "",
+    val id: String
+)
