+++ app/src/main/java/com/senya/novuswidget/ui/home/HomeAction.kt
@@ -1,0 +1,12 @@
+package com.senya.novuswidget.ui.home
+
+import android.net.Uri
+import com.senya.novuswidget.domain.model.ShopItem
+
+sealed interface HomeAction {
+    data class ToggleOpenAddNewCardModal(val isOpen: Boolean) : HomeAction
+    data class AddTempModifiedImage(val uri: Uri?) : HomeAction
+    data class OnChangeCardTitle(val value: String) : HomeAction
+    data class SetModifiedCard(val card: ShopItem) : HomeAction
+    object AddImage : HomeAction
+}
