+++ app/src/main/java/com/senya/novuswidget/ui/home/HomeState.kt
@@ -1,0 +1,10 @@
+package com.senya.novuswidget.ui.home
+
+import com.senya.novuswidget.domain.model.ModifiedShopItem
+import com.senya.novuswidget.domain.model.ShopItem
+
+data class HomeState(
+    val cardList: List<ShopItem> = emptyList(),
+    val isAddNewCardModalOpen: Boolean = false,
+    val modifiedCard: ModifiedShopItem? = null,
+)
